Modifying the Fortran flags in the Makefile
The original Makefile script used nc-config to pull the Fortran library
and include configurations for NetCDF. nc-config provides the C
configuration and then uses nf-config to provide the Fortran
configuration. Unfortunately, the version of NetCDF installed by
Homebrew, both on the Mac and on Linux, does not include support for
nf-config (this issue involves cmake and other details that I care
little about and understand even less).

To deal with this, I worked out the flags that are need, using the
NetCDF website:
https://www.unidata.ucar.edu/software/netcdf/docs/getting_and_building_netcdf.html

For this commit, I implemented the Fortran flags as hard-coded calls,
commenting out the previous calls to nc-config.

	modified:   Makefile